fix upload fail problem

+++ app/src/main/java/com/openpositioning/PositionMe/sensors/Wifi.java
@@ -14,7 +14,7 @@ public class Wifi {
     private String ssid;
     private long bssid;
     private int level;
-    private long frequency;
+    private int frequency;
 
     /**
      * Empty public default constructor of the Wifi object.
@@ -27,7 +27,7 @@ public class Wifi {
     public String getSsid() { return ssid; }
     public long getBssid() { return bssid; }
     public int getLevel() { return level; }
-    public long getFrequency() { return frequency; }
+    public int getFrequency() { return frequency; }
 
     /**
      * Setters for each property
@@ -35,7 +35,7 @@ public class Wifi {
     public void setSsid(String ssid) { this.ssid = ssid; }
     public void setBssid(long bssid) { this.bssid = bssid; }
     public void setLevel(int level) { this.level = level; }
-    public void setFrequency(long frequency) { this.frequency = frequency; }
+    public void setFrequency(int frequency) { this.frequency = frequency; }
 
     /**
      * Generates a string containing mac address and rssi of Wifi.
